django/prodrev_project/requirement.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
Django==6.0
django-allauth==65.13.1
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.3
Flask==3.1.3
flask-cors==6.0.2
fsspec==2026.1.0
groq==1.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.5
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.14.0
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.1
openai==2.32.0
packaging==26.0
pillow==11.3.0
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
setuptools==80.10.2
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.8.3
sqlparse==0.5.5
sympy==1.14.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.1
transformers==5.0.0
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uv==0.9.26
Werkzeug==3.1.8
